requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
azure-storage-blob
pydantic
